♻️ Move singleton to index of styles package
